chore: refactor to separate concerns
