--- packages/extension/src/ui/features/accountActivity/__test__/tokensByNetwork.ts
@@ -1,6 +1,0 @@
-import { Token } from "../../../../shared/token/type"
-import { parsedDefaultTokens } from "../../../../shared/token/utils"
-
-export const tokensByNetwork: Token[] = parsedDefaultTokens.filter(
-  ({ networkId }) => networkId === "goerli-alpha",
-)
