+++ packages/extension/src/ui/features/accountActivity/transform/explorerTransaction/fingerprintExplorerTransaction.ts
@@ -1,10 +1,10 @@
-import { IExplorerTransaction } from "../../../../shared/explorer/type"
-
 /**
  * Derives a simple fingerprint from the event and call names in an Explorer transaction
  *
  * This is used to match known transaction types in the transformer
  */
+
+import { IExplorerTransaction } from "../../../../../shared/explorer/type"
 
 export const fingerprintExplorerTransaction = (
   explorerTransaction: IExplorerTransaction,
