+++ packages/extension/src/ui/features/accountActivity/transform/transaction/__test__/__fixtures__/transaction-calls/goerli-alpha/erc20-swap-my-swap.json
