+++ packages/storybook/src/features/accounts/TransactionCallDataBottomSheet.stories.tsx
@@ -1,4 +1,4 @@
-import { dappAspectBuyNft } from "@argent-x/extension/src/ui/features/accountActivity/__test__/__fixtures__/explorer-transactions/goerli-alpha"
+import { dappAspectBuyNft } from "@argent-x/extension/src/ui/features/accountActivity/transform/explorerTransaction/__test__/__fixtures__/explorer-transactions/goerli-alpha"
 import { TransactionCallDataBottomSheet } from "@argent-x/extension/src/ui/features/accountActivity/ui/TransactionCallDataBottomSheet"
 import { ComponentMeta, ComponentStory } from "@storybook/react"
