+++ packages/storybook/src/features/accounts/TransactionDetailExplorer.stories.tsx
@@ -1,5 +1,6 @@
 import { IExplorerTransaction } from "@argent-x/extension/src/shared/explorer/type"
 import { defaultNetwork } from "@argent-x/extension/src/shared/network"
+import { TransactionDetail } from "@argent-x/extension/src/ui/features/accountActivity/TransactionDetail"
 import {
   accountCreated,
   accountUpgrade,
@@ -15,8 +16,7 @@ import {
   erc20Transfer,
   erc721MintAspect,
   erc721Transfer,
-} from "@argent-x/extension/src/ui/features/accountActivity/__test__/__fixtures__/explorer-transactions/goerli-alpha"
-import { TransactionDetail } from "@argent-x/extension/src/ui/features/accountActivity/TransactionDetail"
+} from "@argent-x/extension/src/ui/features/accountActivity/transform/explorerTransaction/__test__/__fixtures__/explorer-transactions/goerli-alpha"
 import { ComponentMeta, ComponentStory } from "@storybook/react"
 import { MemoryRouter } from "react-router-dom"
